[ENH] DL Refactor (#937)
* switch dl to check_dependences

* switch dl to check_dependences

* make inception time file private

* base class one hot encoder

* switch networks

* inception inport

* regressor refactor

* DL notebook

* DL docs

* examples imports

* examples imports

* examples imports

* refactor tapnet

* example imports

* import Lite

* DL docs
