+++ aeon/classification/deep_learning/_cnn.py
@@ -1,7 +1,7 @@
-"""Multi Layer Perceptron Network (MLP) for classification."""
+"""Time Convolutional Neural Network (CNN) for classification."""
 
-__author__ = ["James-Large", "AurumnPegasus"]
-__all__ = ["MLPClassifier"]
+__author__ = ["James-Large", "TonyBagnall", "hadifawaz1999"]
+__all__ = ["CNNClassifier"]
 
 import gc
 import os
@@ -11,75 +11,93 @@ from copy import deepcopy
 from sklearn.utils import check_random_state
 
 from aeon.classification.deep_learning.base import BaseDeepClassifier
-from aeon.networks.mlp import MLPNetwork
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.networks import CNNNetwork
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
-class MLPClassifier(BaseDeepClassifier):
-    """Multi Layer Perceptron Network (MLP).
+class CNNClassifier(BaseDeepClassifier):
+    """
+    Time Convolutional Neural Network (CNN).
 
     Adapted from the implementation used in [1]_.
 
     Parameters
     ----------
+    n_layers : int, default = 2
+        The number of convolution layers in the network.
+    kernel_size : int or list of int, default = 7
+        Kernel size of convolution layers, if not a list, the same kernel size
+        is used for all layer, len(list) should be n_layers.
+    n_filters : int or list of int, default = [6, 12]
+        Number of filters for each convolution layer, if not a list, the same n_filters
+        is used in all layers.
+    avg_pool_size : int or list of int, default = 3
+        The size of the average pooling layer, if not a list, the same
+        max pooling size is used for all convolution layer.
+    activation : str or list of str, default = "sigmoid"
+        Keras activation function used in the model for each layer, if not a list,
+        the same activation is used for all layers.
+    padding : str or list of str, default = 'valid'
+        The method of padding in convolution layers, if not a list, the same padding
+        used for all convolution layers.
+    strides : int or list of int, default = 1
+        The strides of kernels in the convolution and max pooling layers, if not a
+        list, the same strides are used for all layers.
+    dilation_rate : int or list of int, default = 1
+        The dilation rate of the convolution layers, if not a list, the same dilation
+        rate is used all over the network.
+    use_bias : bool or list of bool, default = True
+        Condition on whether to use bias values for convolution layers,
+        if not a list, the same condition is used for all layers.
+    random_state : int, default = 0
+        Seed to any needed random actions.
     n_epochs : int, default = 2000
-        the number of epochs to train the model
+        The number of epochs to train the model.
     batch_size : int, default = 16
-        the number of samples per gradient update.
-    random_state : int or None, default=None
-        Seed for random number generation.
+        The number of samples per gradient update.
     verbose : boolean, default = False
-        whether to output extra information
-    loss : string, default="mean_squared_error"
-        fit parameter for the keras model
-    file_path : str, default = "./"
-        file_path when saving model_Checkpoint callback
+        Whether to output extra information.
+    loss : string, default = "mean_squared_error"
+        Fit parameter for the keras model.
+    optimizer : keras.optimizer, default = keras.optimizers.Adam()
+    metrics : list of strings, default = ["accuracy"]
+    callbacks : keras.callbacks, default = model_checkpoint
+        To save best model on training loss.
+    file_path : file_path for the best model
+        Only used if checkpoint is used as callback.
     save_best_model : bool, default = False
-        Whether or not to save the best model, if the
-        modelcheckpoint callback is used by default,
-        this condition, if True, will prevent the
-        automatic deletion of the best saved model from
-        file and the user can choose the file name
+        Whether to save the best model, if the modelcheckpoint callback is used by
+        default, this condition, if True, will prevent the automatic deletion of the
+        best saved model from file and the user can choose the file name.
     save_last_model : bool, default = False
-        Whether or not to save the last model, last
-        epoch trained, using the base class method
-        save_last_model_to_file
+        Whether to save the last model, last epoch trained, using the base class method
+        save_last_model_to_file.
     best_file_name : str, default = "best_model"
-        The name of the file of the best model, if
-        save_best_model is set to False, this parameter
-        is discarded
+        The name of the file of the best model, if save_best_model is set to False,
+        this parameter is discarded.
     last_file_name : str, default = "last_model"
-        The name of the file of the last model, if
-        save_last_model is set to False, this parameter
-        is discarded
-    optimizer : keras.optimizer, default=keras.optimizers.Adadelta(),
-    metrics : list of strings, default=["accuracy"],
-    activation : string or a tf callable, default="sigmoid"
-        Activation function used in the output linear layer.
-        List of available activation functions:
-        https://keras.io/api/layers/activations/
-    use_bias : boolean, default = True
-        whether the layer uses a bias vector.
+        The name of the file of the last model, if save_last_model is set to False,
+        this parameter is discarded.
 
     Notes
     -----
-    Adapted from the implementation from source code
-    https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/mlp.py
+    Adapted from the implementation from Fawaz et. al
+    https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/cnn.py
 
     References
     ----------
-    .. [1] Wang et. al, Time series classification from
-    scratch with deep neural networks: A strong baseline,
-    International joint conference on neural networks (IJCNN), 2017.
+    .. [1] Zhao et. al, Convolutional neural networks for time series classification,
+    Journal of Systems Engineering and Electronics, 28(1):2017.
 
     Examples
     --------
-    >>> from aeon.classification.deep_learning.mlp import MLPClassifier
+    >>> from aeon.classification.deep_learning import CNNClassifier
     >>> from aeon.datasets import load_unit_test
     >>> X_train, y_train = load_unit_test(split="train")
-    >>> mlp = MLPClassifier(n_epochs=20, batch_size=4)  # doctest: +SKIP
-    >>> mlp.fit(X_train, y_train)  # doctest: +SKIP
-    MLPClassifier(...)
+    >>> X_test, y_test = load_unit_test(split="test")
+    >>> cnn = CNNClassifier(n_epochs=20, batch_size=4)  # doctest: +SKIP
+    >>> cnn.fit(X_train, y_train)  # doctest: +SKIP
+    CNNClassifier(...)
     """
 
     _tags = {
@@ -90,57 +108,82 @@ class MLPClassifier(BaseDeepClassifier):
 
     def __init__(
         self,
+        n_layers=2,
+        kernel_size=7,
+        n_filters=None,
+        avg_pool_size=3,
+        activation="sigmoid",
+        padding="valid",
+        strides=1,
+        dilation_rate=1,
         n_epochs=2000,
         batch_size=16,
         callbacks=None,
-        verbose=False,
-        loss="categorical_crossentropy",
-        metrics=None,
         file_path="./",
         save_best_model=False,
         save_last_model=False,
         best_file_name="best_model",
         last_file_name="last_model",
+        verbose=False,
+        loss="mean_squared_error",
+        metrics=None,
         random_state=None,
-        activation="sigmoid",
         use_bias=True,
         optimizer=None,
     ):
-        _check_dl_dependencies(severity="error")
-        super(MLPClassifier, self).__init__(last_file_name=last_file_name)
-        self.callbacks = callbacks
-        self.n_epochs = n_epochs
-        self.batch_size = batch_size
-        self.verbose = verbose
-        self.loss = loss
-        self.metrics = metrics
-        self.random_state = random_state
+        _check_soft_dependencies("tensorflow")
+        super(CNNClassifier, self).__init__(last_file_name=last_file_name)
+
+        self.n_layers = n_layers
+        self.kernel_size = kernel_size
+        self.n_filters = n_filters
+        self.padding = padding
+        self.strides = strides
+        self.dilation_rate = dilation_rate
+        self.avg_pool_size = avg_pool_size
         self.activation = activation
         self.use_bias = use_bias
+        self.random_state = random_state
+
+        self.n_epochs = n_epochs
+        self.batch_size = batch_size
+        self.callbacks = callbacks
         self.file_path = file_path
         self.save_best_model = save_best_model
         self.save_last_model = save_last_model
         self.best_file_name = best_file_name
         self.last_file_name = last_file_name
+        self.verbose = verbose
+        self.loss = loss
+        self.metrics = metrics
         self.optimizer = optimizer
         self.history = None
-        self._network = MLPNetwork(
+        self._network = CNNNetwork(
+            n_layers=self.n_layers,
+            kernel_size=self.kernel_size,
+            n_filters=self.n_filters,
+            avg_pool_size=self.avg_pool_size,
+            activation=self.activation,
+            padding=self.padding,
+            strides=self.strides,
+            dilation_rate=self.dilation_rate,
+            use_bias=self.use_bias,
             random_state=self.random_state,
         )
 
     def build_model(self, input_shape, n_classes, **kwargs):
         """Construct a compiled, un-trained, keras model that is ready for training.
 
-        In aeon, time series are stored in numpy arrays of shape (d,m), where d
+        In aeon, time series are stored in numpy arrays of shape (d, m), where d
         is the number of dimensions, m is the series length. Keras/tensorflow assume
-        data is in shape (m,d). This method also assumes (m,d). Transpose should
+        data is in shape (m, d). This method also assumes (m, d). Transpose should
         happen in fit.
 
         Parameters
         ----------
         input_shape : tuple
-            The shape of the data fed into the input layer, should be (m,d)
-        n_classes: int
+            The shape of the data fed into the input layer, should be (m, d)
+        n_classes : int
             The number of classes, which becomes the size of the output layer
 
         Returns
@@ -148,7 +191,6 @@ class MLPClassifier(BaseDeepClassifier):
         output : a compiled Keras Model
         """
         import tensorflow as tf
-        from tensorflow import keras
 
         tf.random.set_seed(self.random_state)
 
@@ -158,20 +200,21 @@ class MLPClassifier(BaseDeepClassifier):
             metrics = self.metrics
         input_layer, output_layer = self._network.build_network(input_shape, **kwargs)
 
-        output_layer = keras.layers.Dense(
+        output_layer = tf.keras.layers.Dense(
             units=n_classes, activation=self.activation, use_bias=self.use_bias
         )(output_layer)
 
         self.optimizer_ = (
-            keras.optimizers.Adadelta() if self.optimizer is None else self.optimizer
+            tf.keras.optimizers.Adam() if self.optimizer is None else self.optimizer
         )
 
-        model = keras.models.Model(inputs=input_layer, outputs=output_layer)
+        model = tf.keras.models.Model(inputs=input_layer, outputs=output_layer)
         model.compile(
             loss=self.loss,
             optimizer=self.optimizer_,
             metrics=metrics,
         )
+
         return model
 
     def _fit(self, X, y):
@@ -208,9 +251,6 @@ class MLPClassifier(BaseDeepClassifier):
 
         self.callbacks_ = (
             [
-                tf.keras.callbacks.ReduceLROnPlateau(
-                    monitor="loss", factor=0.5, patience=200, min_lr=0.1
-                ),
                 tf.keras.callbacks.ModelCheckpoint(
                     filepath=self.file_path + self.file_name_ + ".hdf5",
                     monitor="loss",
@@ -251,9 +291,9 @@ class MLPClassifier(BaseDeepClassifier):
 
         Parameters
         ----------
-        parameter_set : str, default="default"
+        parameter_set : str, default = "default"
             Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
+            special parameters are defined for a value, will return "default" set.
             For classifiers, a "default" set of parameters should be provided for
             general testing, and a "results_comparison" set for comparing against
             previously recorded results if the general set does not produce suitable
@@ -261,7 +301,7 @@ class MLPClassifier(BaseDeepClassifier):
 
         Returns
         -------
-        params : dict or list of dict, default={}
+        params : dict or list of dict, default = {}
             Parameters to create testing instances of the class.
             Each dict are parameters to construct an "interesting" test instance, i.e.,
             `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
@@ -270,7 +310,7 @@ class MLPClassifier(BaseDeepClassifier):
         param1 = {
             "n_epochs": 10,
             "batch_size": 4,
-            "use_bias": False,
+            "avg_pool_size": 4,
         }
 
         test_params = [param1]
