+++ aeon/classification/deep_learning/_fcn.py
@@ -1,7 +1,7 @@
-"""Encoder Classifier."""
+"""Fully Convolutional Network (FCN) for classification."""
 
-__author__ = ["hadifawaz1999"]
-__all__ = ["EncoderClassifier"]
+__author__ = ["James-Large", "AurumnPegasus", "hadifawaz1999"]
+__all__ = ["FCNClassifier"]
 
 import gc
 import os
@@ -11,38 +11,50 @@ from copy import deepcopy
 from sklearn.utils import check_random_state
 
 from aeon.classification.deep_learning.base import BaseDeepClassifier
-from aeon.networks.encoder import EncoderNetwork
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.networks import FCNNetwork
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
-class EncoderClassifier(BaseDeepClassifier):
-    """
-    Establish the network structure for an Encoder.
+class FCNClassifier(BaseDeepClassifier):
+    """Fully Convolutional Network (FCN).
 
     Adapted from the implementation used in [1]_.
 
     Parameters
     ----------
-    kernel_size : array of int, default = [5, 11, 21]
-        Specifying the length of the 1D convolution windows.
-    n_filters : array of int, default = [128, 256, 512]
-        Specifying the number of 1D convolution filters used for each layer,
-        the shape of this array should be the same as kernel_size.
-    max_pool_size : int, default = 2
-        Size of the max pooling windows.
-    activation : string, default = sigmoid
-        Keras activation function.
-    dropout_proba : float, default = 0.2
-        Specifying the dropout layer probability.
-    padding : string, default = same
-        Specifying the type of padding used for the 1D convolution.
-    strides : int, default = 1
-        Specifying the sliding rate of the 1D convolution filter.
-    fc_units : int, default = 256
-        Specifying the number of units in the hidden fully
-        connected layer used in the EncoderNetwork.
+    n_layers : int, default = 3
+        Number of convolution layers.
+    n_filters : int or list of int, default = [128,256,128]
+        Number of filters used in convolution layers.
+    kernel_size : int or list of int, default = [8,5,3]
+        Size of convolution kernel.
+    dilation_rate : int or list of int, default = 1
+        The dilation rate for convolution.
+    strides : int or list of int, default = 1
+        The strides of the convolution filter.
+    padding : str or list of str, default = "same"
+        The type of padding used for convolution.
+    activation : str or list of str, default = "relu"
+        Activation used after the convolution.
+    use_bias : bool or list of bool, default = True
+        Whether or not ot use bias in convolution.
+    n_epochs : int, default = 2000
+        The number of epochs to train the model.
+    batch_size : int, default = 16
+        The number of samples per gradient update.
+    use_mini_batch_size : bool, default = True
+        Whether or not to use the mini batch size formula.
+    random_state : int or None, default = None
+        Seed for random number generation.
+    verbose : boolean, default = False
+        Whether to output extra information.
+    loss : string, default = "mean_squared_error"
+        Fit parameter for the keras model.
+    metrics : list of strings, default = ["accuracy"]
+    optimizer : keras.optimizers object, default = Adam(lr=0.01)
+        Specify the optimizer and the learning rate to be used.
     file_path : str, default = "./"
-        File path when saving model_Checkpoint callback.
+        File path to save best model.
     save_best_model : bool, default = False
         Whether or not to save the best model, if the
         modelcheckpoint callback is used by default,
@@ -61,47 +73,52 @@ class EncoderClassifier(BaseDeepClassifier):
         The name of the file of the last model, if
         save_last_model is set to False, this parameter
         is discarded.
-    random_state : int, default = 0
-        Seed to any needed random actions.
-
+    callbacks : keras.callbacks, default = None
     Notes
     -----
-    Adapted from source code
-    https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/encoder.py
+    Adapted from the implementation from Fawaz et. al
+    https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/fcn.py
 
     References
     ----------
-    .. [1] Serrà et al. Towards a Universal Neural Network Encoder for Time Series
-    In proceedings International Conference of the Catalan Association
-    for Artificial Intelligence, 120--129 2018.
+    .. [1] Zhao et. al, Convolutional neural networks for time series classification,
+    Journal of Systems Engineering and Electronics, 28(1):2017.
 
-
+    Examples
+    --------
+    >>> from aeon.classification.deep_learning import FCNClassifier
+    >>> from aeon.datasets import load_unit_test
+    >>> X_train, y_train = load_unit_test(split="train", return_X_y=True)
+    >>> X_test, y_test = load_unit_test(split="test", return_X_y=True)
+    >>> fcn = FCNClassifier(n_epochs=20, batch_size=4)  # doctest: +SKIP
+    >>> fcn.fit(X_train, y_train)  # doctest: +SKIP
+    FCNClassifier(...)
     """
 
     _tags = {
-        "python_dependencies": ["tensorflow", "tensorflow_addons"],
+        "python_dependencies": "tensorflow",
         "capability:multivariate": True,
         "algorithm_type": "deeplearning",
     }
 
     def __init__(
         self,
-        n_epochs=100,
-        batch_size=12,
-        kernel_size=None,
+        n_layers=3,
         n_filters=None,
-        dropout_proba=0.2,
-        activation="sigmoid",
-        max_pool_size=2,
-        padding="same",
+        kernel_size=None,
+        dilation_rate=1,
         strides=1,
-        fc_units=256,
-        callbacks=None,
+        padding="same",
+        activation="relu",
         file_path="./",
         save_best_model=False,
         save_last_model=False,
         best_file_name="best_model",
         last_file_name="last_model",
+        n_epochs=2000,
+        batch_size=16,
+        use_mini_batch_size=True,
+        callbacks=None,
         verbose=False,
         loss="categorical_crossentropy",
         metrics=None,
@@ -109,52 +126,51 @@ class EncoderClassifier(BaseDeepClassifier):
         use_bias=True,
         optimizer=None,
     ):
-        _check_dl_dependencies(severity="error")
-        super(EncoderClassifier, self).__init__(last_file_name=last_file_name)
+        _check_soft_dependencies("tensorflow")
+        super(FCNClassifier, self).__init__(last_file_name=last_file_name)
 
-        self.n_filters = n_filters
-        self.max_pool_size = max_pool_size
+        self.n_layers = n_layers
         self.kernel_size = kernel_size
+        self.n_filters = n_filters
         self.strides = strides
         self.activation = activation
+        self.dilation_rate = dilation_rate
         self.padding = padding
-        self.dropout_proba = dropout_proba
-        self.fc_units = fc_units
+        self.use_bias = use_bias
 
         self.callbacks = callbacks
+        self.n_epochs = n_epochs
+        self.batch_size = batch_size
+        self.use_mini_batch_size = use_mini_batch_size
+        self.verbose = verbose
+        self.loss = loss
+        self.metrics = metrics
+        self.random_state = random_state
+        self.optimizer = optimizer
+        self.history = None
         self.file_path = file_path
         self.save_best_model = save_best_model
         self.save_last_model = save_last_model
         self.best_file_name = best_file_name
         self.last_file_name = last_file_name
-        self.n_epochs = n_epochs
-        self.batch_size = batch_size
-        self.verbose = verbose
-        self.loss = loss
-        self.metrics = metrics
-        self.random_state = random_state
-        self.use_bias = use_bias
-        self.optimizer = optimizer
-        self.history = None
-
-        self._network = EncoderNetwork(
+        self._network = FCNNetwork(
+            random_state=self.random_state,
+            n_layers=self.n_layers,
             kernel_size=self.kernel_size,
-            max_pool_size=self.max_pool_size,
             n_filters=self.n_filters,
-            fc_units=self.fc_units,
             strides=self.strides,
             padding=self.padding,
-            dropout_proba=self.dropout_proba,
+            dilation_rate=self.dilation_rate,
             activation=self.activation,
-            random_state=self.random_state,
+            use_bias=self.use_bias,
         )
 
     def build_model(self, input_shape, n_classes, **kwargs):
         """Construct a compiled, un-trained, keras model that is ready for training.
 
-        In aeon, time series are stored in numpy arrays of shape (d, m), where d
+        In aeon, time series are stored in numpy arrays of shape (d,m), where d
         is the number of dimensions, m is the series length. Keras/tensorflow assume
-        data is in shape (m, d). This method also assumes (m, d). Transpose should
+        data is in shape (m,d). This method also assumes (m,d). Transpose should
         happen in fit.
 
         Parameters
@@ -179,13 +195,11 @@ class EncoderClassifier(BaseDeepClassifier):
         input_layer, output_layer = self._network.build_network(input_shape, **kwargs)
 
         output_layer = tf.keras.layers.Dense(
-            units=n_classes, activation=self.activation, use_bias=self.use_bias
+            units=n_classes, activation="softmax", use_bias=self.use_bias
         )(output_layer)
 
         self.optimizer_ = (
-            tf.keras.optimizers.Adam(learning_rate=0.00001)
-            if self.optimizer is None
-            else self.optimizer
+            tf.keras.optimizers.Adam() if self.optimizer is None else self.optimizer
         )
 
         model = tf.keras.models.Model(inputs=input_layer, outputs=output_layer)
@@ -225,12 +239,20 @@ class EncoderClassifier(BaseDeepClassifier):
         if self.verbose:
             self.training_model_.summary()
 
+        if self.use_mini_batch_size:
+            mini_batch_size = min(self.batch_size, X.shape[0] // 10)
+        else:
+            mini_batch_size = self.batch_size
+
         self.file_name_ = (
             self.best_file_name if self.save_best_model else str(time.time_ns())
         )
 
         self.callbacks_ = (
             [
+                tf.keras.callbacks.ReduceLROnPlateau(
+                    monitor="loss", factor=0.5, patience=50, min_lr=0.0001
+                ),
                 tf.keras.callbacks.ModelCheckpoint(
                     filepath=self.file_path + self.file_name_ + ".hdf5",
                     monitor="loss",
@@ -244,7 +266,7 @@ class EncoderClassifier(BaseDeepClassifier):
         self.history = self.training_model_.fit(
             X,
             y_onehot,
-            batch_size=self.batch_size,
+            batch_size=mini_batch_size,
             epochs=self.n_epochs,
             verbose=self.verbose,
             callbacks=self.callbacks_,
@@ -288,12 +310,12 @@ class EncoderClassifier(BaseDeepClassifier):
             `create_test_instance` uses the first (or only) dictionary in `params`.
         """
         param1 = {
-            "n_epochs": 8,
+            "n_epochs": 10,
             "batch_size": 4,
             "use_bias": False,
-            "fc_units": 8,
+            "n_layers": 2,
+            "padding": "valid",
             "strides": 2,
-            "dropout_proba": 0,
         }
 
         test_params = [param1]
