+++ aeon/classification/deep_learning/_tapnet.py
@@ -2,9 +2,11 @@
 
 __author__ = [
     "Jack Russon",
+    "TonyBagnall",
+    "achieveordie",
 ]
 __all__ = [
-    "TapNetRegressor",
+    "TapNetClassifier",
 ]
 
 import gc
@@ -12,72 +14,77 @@ from copy import deepcopy
 
 from sklearn.utils import check_random_state
 
-from aeon.networks.tapnet import TapNetNetwork
-from aeon.regression.deep_learning.base import BaseDeepRegressor
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.classification.deep_learning.base import BaseDeepClassifier
+from aeon.networks import TapNetNetwork
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
-class TapNetRegressor(BaseDeepRegressor):
+class TapNetClassifier(BaseDeepClassifier):
     """Time series attentional prototype network (TapNet).
 
-    Adapted from the implementation used in [1]_.
-
-    TapNet was initially proposed for multivariate time series classification. The is an
-    adaptation for time series regression. TapNet comprises these components: random
-    dimension permutation, multivariate time series encoding, and attentional
-    prototype learning.
+     Implemented as described in [1]_.
 
     Parameters
     ----------
-    filter_sizes        : array of int, default = (256, 256, 128)
+    filter_sizes : array of int, default = (256, 256, 128)
         sets the kernel size argument for each convolutional block.
         Controls number of convolutional filters
         and number of neurons in attention dense layers.
-    kernel_size        : array of int, default = (8, 5, 3)
+    kernel_sizes : array of int, default = (8, 5, 3)
         controls the size of the convolutional kernels
-    layers              : array of int, default = (500, 300)
+    layers : array of int, default = (500, 300)
         size of dense layers
-    n_epochs            : int, default = 2000
+    reduction : int, default = 16
+        divides the number of dense neurons in the first layer of the attention block.
+    n_epochs : int, default = 2000
         number of epochs to train the model
-    batch_size          : int, default = 16
+    batch_size : int, default = 16
         number of samples per update
-    dropout             : float, default = 0.5
+    dropout : float, default = 0.5
         dropout rate, in the range [0, 1)
-    dilation            : int, default = 1
+    dilation : int, default = 1
         dilation value
-    activation          : str, default = "sigmoid"
+    activation : str, default = "sigmoid"
         activation function for the last output layer
-    loss                : str, default = "mean_squared_error"
+    loss : str, default = "binary_crossentropy"
         loss function for the classifier
-    optimizer           : str or None, default = "Adam(lr=0.01)"
-        gradient updating function for the classifier
-    use_bias            : bool, default = True
+    optimizer : str or None, default = "Adam(lr=0.01)"
+        gradient updating function for the classifer
+    use_bias : bool, default = True
         whether to use bias in the output dense layer
-    use_rp              : bool, default = True
+    use_rp : bool, default = True
         whether to use random projections
-    use_att             : bool, default = True
+    use_att : bool, default = True
         whether to use self attention
-    use_lstm        : bool, default = True
+    use_lstm : bool, default = True
         whether to use an LSTM layer
-    use_cnn         : bool, default = True
+    use_cnn : bool, default = True
         whether to use a CNN layer
-    verbose         : bool, default = False
+    verbose : bool, default = False
         whether to output extra information
-    random_state    : int or None, default = None
+    random_state : int or None, default = None
         seed for random
+
+    Attributes
+    ----------
+    n_classes      : int
+        number of classes extracted from the data
 
     References
     ----------
     .. [1] Zhang et al. Tapnet: Multivariate time series classification with
-    attentional prototypical network,
-    Proceedings of the AAAI Conference on Artificial Intelligence
-    34(4), 6845-6852, 2020
+    attentional prototypical network, Proceedings of the AAAI Conference on
+    Artificial Intelligence 34(4), 6845-6852, 2020
 
-    Notes
-    -----
-    The Implementation of TapNet found at https://github.com/kdd2019-tapnet/tapnet
-    Currently does not implement custom distance matrix loss function
-    or class  based self attention.
+    Examples
+    --------
+    >>> from aeon.classification.deep_learning import TapNetClassifier
+    >>> from aeon.datasets import load_unit_test
+    >>> X_train, y_train = load_unit_test(split="train")
+    >>> X_test, y_test = load_unit_test(split="test")
+    >>> tapnet = TapNetClassifier(n_epochs=20,batch_size=4)  # doctest: +SKIP
+    >>> tapnet.fit(X_train, y_train) # doctest: +SKIP
+    TapNetClassifier(...)
     """
 
     _tags = {"python_dependencies": "tensorflow"}
@@ -92,22 +99,22 @@ class TapNetRegressor(BaseDeepRegressor):
         dilation=1,
         layers=(500, 300),
         use_rp=True,
-        activation=None,
         rp_params=(-1, 3),
+        activation="sigmoid",
         use_bias=True,
         use_att=True,
         use_lstm=True,
         use_cnn=True,
         random_state=None,
         padding="same",
-        loss="mean_squared_error",
+        loss="binary_crossentropy",
         optimizer=None,
         metrics=None,
         callbacks=None,
         verbose=False,
     ):
-        _check_dl_dependencies(severity="error")
-        super(TapNetRegressor, self).__init__()
+        _check_soft_dependencies("tensorflow")
+        super(TapNetClassifier, self).__init__()
 
         self.batch_size = batch_size
         self.random_state = random_state
@@ -151,7 +158,7 @@ class TapNetRegressor(BaseDeepRegressor):
             padding=self.padding,
         )
 
-    def build_model(self, input_shape, **kwargs):
+    def build_model(self, input_shape, n_classes, **kwargs):
         """Construct a complied, un-trained, keras model that is ready for training.
 
         In aeon, time series are stored in numpy arrays of shape (d,m), where d
@@ -163,6 +170,8 @@ class TapNetRegressor(BaseDeepRegressor):
         ----------
         input_shape     : tuple
             The shape of the data fed into the input layer, should be (m, d)
+        n_classes       : int
+            The number of classes, which becomes the size of the output layer
 
         Returns
         -------
@@ -173,12 +182,15 @@ class TapNetRegressor(BaseDeepRegressor):
 
         tf.random.set_seed(self.random_state)
 
-        metrics = ["mean_squared_error"] if self.metrics is None else self.metrics
+        if self.metrics is None:
+            metrics = ["accuracy"]
+        else:
+            metrics = self.metrics
 
         input_layer, output_layer = self._network.build_network(input_shape, **kwargs)
 
         output_layer = keras.layers.Dense(
-            units=1, activation=self.activation, use_bias=self.use_bias
+            units=n_classes, activation=self.activation, use_bias=self.use_bias
         )(output_layer)
 
         self.optimizer_ = (
@@ -197,32 +209,31 @@ class TapNetRegressor(BaseDeepRegressor):
         return model
 
     def _fit(self, X, y):
-        """
-        Fit the regressor on the training set (X, y).
+        """Fit the classifier on the training set (X, y).
 
         Parameters
         ----------
-        X   : np.ndarray of shape = (n_instances(n), n_channels(d), series_length(m))
+        X : np.ndarray of shape = (n_instances(n), n_channels(d), series_length(m))
             Input training samples
-        y   : np.ndarray of shape n
-            Input training responses
+        y : np.ndarray of shape n
+            Input training class labels
 
         Returns
         -------
         self: object
         """
-        # Transpose to conform to expectation format from keras
+        y_onehot = self.convert_y_to_keras(y)
+        # Transpose to conform to expectation format by keras
         X = X.transpose(0, 2, 1)
 
         check_random_state(self.random_state)
         self.input_shape = X.shape[1:]
-
-        self.model_ = self.build_model(self.input_shape)
+        self.model_ = self.build_model(self.input_shape, self.n_classes_)
         if self.verbose:
             self.model_.summary()
         self.history = self.model_.fit(
             X,
-            y,
+            y_onehot,
             batch_size=self.batch_size,
             epochs=self.n_epochs,
             verbose=self.verbose,
@@ -240,11 +251,7 @@ class TapNetRegressor(BaseDeepRegressor):
         ----------
         parameter_set : str, default="default"
             Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
-            For classifiers, a "default" set of parameters should be provided for
-            general testing, and a "results_comparison" set for comparing against
-            previously recorded results if the general set does not produce suitable
-            probabilities to compare against.
+            special parameters are defined for a value, will return "default" set.
 
         Returns
         -------
@@ -255,11 +262,12 @@ class TapNetRegressor(BaseDeepRegressor):
             `create_test_instance` uses the first (or only) dictionary in `params`.
         """
         param1 = {
-            "n_epochs": 10,
+            "n_epochs": 20,
             "batch_size": 4,
-            "padding": "valid",
+            "use_lstm": False,
+            "use_att": False,
             "filter_sizes": (16, 16, 16),
-            "kernel_size": (3, 3, 1),
-            "layers": (25, 50),
+            "dilation": 2,
+            "layers": (32, 16),
         }
         return [param1]
