+++ aeon/clustering/deep_learning/_ae_fcn.py
@@ -1,7 +1,6 @@
-"""Fully Convolutional Network (FCN) for classification."""
-
-__author__ = ["James-Large", "AurumnPegasus", "hadifawaz1999"]
-__all__ = ["FCNClassifier"]
+"""Deep Learning Auto-Encoder using FCN Network."""
+__author__ = ["hadifawaz1999"]
+__all__ = ["AEFCNClusterer"]
 
 import gc
 import os
@@ -10,18 +9,26 @@ from copy import deepcopy
 
 from sklearn.utils import check_random_state
 
-from aeon.classification.deep_learning.base import BaseDeepClassifier
-from aeon.networks.fcn import FCNNetwork
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.clustering.deep_learning.base import BaseDeepClusterer
+from aeon.networks import AEFCNNetwork
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
-class FCNClassifier(BaseDeepClassifier):
-    """Fully Convolutional Network (FCN).
-
-    Adapted from the implementation used in [1]_.
+class AEFCNClusterer(BaseDeepClusterer):
+    """Auto-Encoder based Fully Convolutional Network (FCN), as described in [1]_.
 
     Parameters
     ----------
+    n_clusters : int, default=None
+        Number of clusters for the deep learnign model.
+    clustering_algorithm : str, default="kmeans"
+        The clustering algorithm used in the latent space.
+    clustering_params : dict, default=None
+        Dictionary containing the parameters of the clustering algorithm chosen.
+    latent_space_dim : int, default=128
+        Dimension of the latent space of the auto-encoder.
+    temporal_latent_space : bool, default = False
+        Flag to choose whether the latent space is an MTS or Euclidean space.
     n_layers : int, default = 3
         Number of convolution layers.
     n_filters : int or list of int, default = [128,256,128]
@@ -42,15 +49,14 @@ class FCNClassifier(BaseDeepClassifier):
         The number of epochs to train the model.
     batch_size : int, default = 16
         The number of samples per gradient update.
-    use_mini_batch_size : bool, default = True
+    use_mini_batch_size : bool, default = True,
         Whether or not to use the mini batch size formula.
-    random_state : int or None, default = None
+    random_state : int or None, default=None
         Seed for random number generation.
     verbose : boolean, default = False
         Whether to output extra information.
-    loss : string, default = "mean_squared_error"
+    loss : string, default="mean_squared_error"
         Fit parameter for the keras model.
-    metrics : list of strings, default = ["accuracy"]
     optimizer : keras.optimizers object, default = Adam(lr=0.01)
         Specify the optimizer and the learning rate to be used.
     file_path : str, default = "./"
@@ -74,6 +80,8 @@ class FCNClassifier(BaseDeepClassifier):
         save_last_model is set to False, this parameter
         is discarded.
     callbacks : keras.callbacks, default = None
+        List of keras callbacks.
+
     Notes
     -----
     Adapted from the implementation from Fawaz et. al
@@ -86,13 +94,13 @@ class FCNClassifier(BaseDeepClassifier):
 
     Examples
     --------
-    >>> from aeon.classification.deep_learning.fcn import FCNClassifier
+    >>> from aeon.clustering.deep_learning import AEFCNClusterer
     >>> from aeon.datasets import load_unit_test
     >>> X_train, y_train = load_unit_test(split="train", return_X_y=True)
     >>> X_test, y_test = load_unit_test(split="test", return_X_y=True)
-    >>> fcn = FCNClassifier(n_epochs=20, batch_size=4)  # doctest: +SKIP
-    >>> fcn.fit(X_train, y_train)  # doctest: +SKIP
-    FCNClassifier(...)
+    >>> aefcn = AEFCNClusterer(n_clusters=2,n_epochs=20,batch_size=4)  # doctest: +SKIP
+    >>> aefcn.fit(X_train)  # doctest: +SKIP
+    AEFCNClusterer(...)
     """
 
     _tags = {
@@ -103,6 +111,11 @@ class FCNClassifier(BaseDeepClassifier):
 
     def __init__(
         self,
+        n_clusters,
+        clustering_algorithm="kmeans",
+        clustering_params=None,
+        latent_space_dim=128,
+        temporal_latent_space=False,
         n_layers=3,
         n_filters=None,
         kernel_size=None,
@@ -110,116 +123,118 @@ class FCNClassifier(BaseDeepClassifier):
         strides=1,
         padding="same",
         activation="relu",
+        use_bias=True,
+        n_epochs=2000,
+        batch_size=32,
+        use_mini_batch_size=False,
+        random_state=0,
+        verbose=False,
+        loss="mse",
+        optimizer="Adam",
         file_path="./",
         save_best_model=False,
         save_last_model=False,
         best_file_name="best_model",
-        last_file_name="last_model",
-        n_epochs=2000,
-        batch_size=16,
-        use_mini_batch_size=True,
+        last_file_name="last_file",
         callbacks=None,
-        verbose=False,
-        loss="categorical_crossentropy",
-        metrics=None,
-        random_state=None,
-        use_bias=True,
-        optimizer=None,
     ):
-        _check_dl_dependencies(severity="error")
-        super(FCNClassifier, self).__init__(last_file_name=last_file_name)
-
-        self.n_layers = n_layers
-        self.kernel_size = kernel_size
-        self.n_filters = n_filters
-        self.strides = strides
-        self.activation = activation
-        self.dilation_rate = dilation_rate
-        self.padding = padding
-        self.use_bias = use_bias
-
-        self.callbacks = callbacks
-        self.n_epochs = n_epochs
+        _check_soft_dependencies("tensorflow")
+        super(AEFCNClusterer, self).__init__(
+            n_clusters=n_clusters,
+            clustering_algorithm=clustering_algorithm,
+            clustering_params=clustering_params,
+            batch_size=batch_size,
+            last_file_name=last_file_name,
+        )
+        self.n_clusters = n_clusters
+        self.clustering_algorithm = clustering_algorithm
+        self.clustering_params = clustering_params
         self.batch_size = batch_size
-        self.use_mini_batch_size = use_mini_batch_size
-        self.verbose = verbose
-        self.loss = loss
-        self.metrics = metrics
-        self.random_state = random_state
+        self.last_file_name = last_file_name
+        self.latent_space_dim = latent_space_dim
+        self.temporal_latent_space = temporal_latent_space
+        self.n_layers = n_layers
+        self.n_filters = n_filters
+        self.kernel_size = kernel_size
+        self.activation = activation
+        self.padding = padding
+        self.strides = strides
+        self.dilation_rate = dilation_rate
+        self.use_bias = use_bias
         self.optimizer = optimizer
-        self.history = None
+        self.loss = loss
+        self.verbose = verbose
+        self.use_mini_batch_size = use_mini_batch_size
+        self.callbacks = callbacks
         self.file_path = file_path
+        self.n_epochs = n_epochs
         self.save_best_model = save_best_model
         self.save_last_model = save_last_model
         self.best_file_name = best_file_name
-        self.last_file_name = last_file_name
-        self._network = FCNNetwork(
-            random_state=self.random_state,
+        self.random_state = random_state
+
+        self._network = AEFCNNetwork(
+            latent_space_dim=self.latent_space_dim,
+            temporal_latent_space=self.temporal_latent_space,
             n_layers=self.n_layers,
-            kernel_size=self.kernel_size,
             n_filters=self.n_filters,
+            kernel_size=self.kernel_size,
+            dilation_rate=self.dilation_rate,
             strides=self.strides,
             padding=self.padding,
-            dilation_rate=self.dilation_rate,
             activation=self.activation,
             use_bias=self.use_bias,
+            random_state=self.random_state,
         )
 
-    def build_model(self, input_shape, n_classes, **kwargs):
+    def build_model(self, input_shape, **kwargs):
         """Construct a compiled, un-trained, keras model that is ready for training.
 
-        In aeon, time series are stored in numpy arrays of shape (d,m), where d
-        is the number of dimensions, m is the series length. Keras/tensorflow assume
-        data is in shape (m,d). This method also assumes (m,d). Transpose should
-        happen in fit.
+        In aeon, time series are stored in numpy arrays of shape
+        (n_channels,n_timepoints). Keras/tensorflow assume
+        data is in shape (n_timepoints,n_channels). This method also assumes
+        (n_timepoints,n_channels). Transpose should happen in fit.
 
         Parameters
         ----------
         input_shape : tuple
-            The shape of the data fed into the input layer, should be (m, d).
-        n_classes : int
-            The number of classes, which becomes the size of the output layer.
+            The shape of the data fed into the input layer, should be
+            (n_timepoints,n_channels).
 
         Returns
         -------
-        output : a compiled Keras Model
+        output : a compiled Keras Model.
         """
         import tensorflow as tf
 
         tf.random.set_seed(self.random_state)
 
-        if self.metrics is None:
-            metrics = ["accuracy"]
-        else:
-            metrics = self.metrics
-        input_layer, output_layer = self._network.build_network(input_shape, **kwargs)
+        encoder, decoder = self._network.build_network(input_shape, **kwargs)
 
-        output_layer = tf.keras.layers.Dense(
-            units=n_classes, activation="softmax", use_bias=self.use_bias
-        )(output_layer)
+        input_layer = tf.keras.layers.Input(input_shape, name="input layer")
+        encoder_output = encoder(input_layer)
+        decoder_output = decoder(encoder_output)
+        output_layer = tf.keras.layers.Reshape(
+            target_shape=input_shape, name="outputlayer"
+        )(decoder_output)
+
+        model = tf.keras.models.Model(inputs=input_layer, outputs=output_layer)
 
         self.optimizer_ = (
             tf.keras.optimizers.Adam() if self.optimizer is None else self.optimizer
         )
 
-        model = tf.keras.models.Model(inputs=input_layer, outputs=output_layer)
-        model.compile(
-            loss=self.loss,
-            optimizer=self.optimizer_,
-            metrics=metrics,
-        )
+        model.compile(optimizer=self.optimizer_, loss=self.loss)
 
         return model
 
-    def _fit(self, X, y):
+    def _fit(self, X):
         """Fit the classifier on the training set (X, y).
 
         Parameters
         ----------
-        X : np.ndarray of shape = (n_instances (n), n_channels (d), series_length (m))
+        X : np.ndarray of shape = (n_instances (n), n_channels (d), n_timepoints (m))
             The training input samples.
-        y : np.ndarray of shape n
-            The training data class labels.
 
         Returns
         -------
@@ -227,14 +242,13 @@ class FCNClassifier(BaseDeepClassifier):
         """
         import tensorflow as tf
 
-        y_onehot = self.convert_y_to_keras(y)
         # Transpose to conform to Keras input style.
         X = X.transpose(0, 2, 1)
 
         check_random_state(self.random_state)
 
         self.input_shape = X.shape[1:]
-        self.training_model_ = self.build_model(self.input_shape, self.n_classes_)
+        self.training_model_ = self.build_model(self.input_shape)
 
         if self.verbose:
             self.training_model_.summary()
@@ -265,7 +279,7 @@ class FCNClassifier(BaseDeepClassifier):
 
         self.history = self.training_model_.fit(
             X,
-            y_onehot,
+            X,
             batch_size=mini_batch_size,
             epochs=self.n_epochs,
             verbose=self.verbose,
@@ -281,11 +295,17 @@ class FCNClassifier(BaseDeepClassifier):
         except FileNotFoundError:
             self.model_ = deepcopy(self.training_model_)
 
-        if self.save_last_model:
-            self.save_last_model_to_file(file_path=self.file_path)
+        self._fit_clustering(X=X)
 
         gc.collect()
+
         return self
+
+    def _score(self, X, y=None):
+        # Transpose to conform to Keras input style.
+        X = X.transpose(0, 2, 1)
+        latent_space = self.model_.layers[1].predict(X)
+        return self.clusterer.score(latent_space)
 
     @classmethod
     def get_test_params(cls, parameter_set="default"):
@@ -293,9 +313,9 @@ class FCNClassifier(BaseDeepClassifier):
 
         Parameters
         ----------
-        parameter_set : str, default = "default"
+        parameter_set : str, default="default"
             Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return "default" set.
+            special parameters are defined for a value, will return `"default"` set.
             For classifiers, a "default" set of parameters should be provided for
             general testing, and a "results_comparison" set for comparing against
             previously recorded results if the general set does not produce suitable
@@ -303,19 +323,26 @@ class FCNClassifier(BaseDeepClassifier):
 
         Returns
         -------
-        params : dict or list of dict, default = {}
+        params : dict or list of dict, default={}
             Parameters to create testing instances of the class.
             Each dict are parameters to construct an "interesting" test instance, i.e.,
             `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
             `create_test_instance` uses the first (or only) dictionary in `params`.
         """
         param1 = {
-            "n_epochs": 10,
+            "n_clusters": 2,
+            "n_epochs": 1,
             "batch_size": 4,
             "use_bias": False,
-            "n_layers": 2,
-            "padding": "valid",
-            "strides": 2,
+            "n_layers": 1,
+            "padding": "same",
+            "strides": 1,
+            "clustering_params": {
+                "distance": "euclidean",
+                "averaging_method": "mean",
+                "n_init": 1,
+                "max_iter": 30,
+            },
         }
 
         test_params = [param1]
