+++ aeon/networks/__init__.py
@@ -9,13 +9,15 @@ __all__ = [
     "ResNetNetwork",
     "TapNetNetwork",
     "AEFCNNetwork",
+    "LITENetwork",
 ]
-from aeon.networks.ae_fcn import AEFCNNetwork
+from aeon.networks._ae_fcn import AEFCNNetwork
+from aeon.networks._cnn import CNNNetwork
+from aeon.networks._encoder import EncoderNetwork
+from aeon.networks._fcn import FCNNetwork
+from aeon.networks._inception import InceptionNetwork
+from aeon.networks._lite import LITENetwork
+from aeon.networks._mlp import MLPNetwork
+from aeon.networks._resnet import ResNetNetwork
+from aeon.networks._tapnet import TapNetNetwork
 from aeon.networks.base import BaseDeepNetwork
-from aeon.networks.cnn import CNNNetwork
-from aeon.networks.encoder import EncoderNetwork
-from aeon.networks.fcn import FCNNetwork
-from aeon.networks.inception import InceptionNetwork
-from aeon.networks.mlp import MLPNetwork
-from aeon.networks.resnet import ResNetNetwork
-from aeon.networks.tapnet import TapNetNetwork
