+++ aeon/networks/_lite.py
@@ -2,7 +2,7 @@
 __author__ = ["hadifawaz1999"]
 
 from aeon.networks.base import BaseDeepNetwork
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
 class LITENetwork(BaseDeepNetwork):
@@ -46,7 +46,7 @@ class LITENetwork(BaseDeepNetwork):
         activation="relu",
         random_state=0,
     ):
-        _check_dl_dependencies(severity="error")
+        _check_soft_dependencies("tensorflow")
 
         self.nb_filters = nb_filters
         self.kernel_size = kernel_size
