+++ aeon/regression/deep_learning/__init__.py
@@ -8,11 +8,11 @@ __all__ = [
     "TapNetRegressor",
 ]
 
-from aeon.regression.deep_learning.cnn import CNNRegressor
-from aeon.regression.deep_learning.fcn import FCNRegressor
-from aeon.regression.deep_learning.inception_time import (
+from aeon.regression.deep_learning._cnn import CNNRegressor
+from aeon.regression.deep_learning._fcn import FCNRegressor
+from aeon.regression.deep_learning._inception_time import (
     InceptionTimeRegressor,
     IndividualInceptionRegressor,
 )
-from aeon.regression.deep_learning.resnet import ResNetRegressor
-from aeon.regression.deep_learning.tapnet import TapNetRegressor
+from aeon.regression.deep_learning._resnet import ResNetRegressor
+from aeon.regression.deep_learning._tapnet import TapNetRegressor
